move away from flat file structure
