+++ src/NewsletterRoute.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 
-import NewsletterView from './NewsletterView'
+import NewsletterView from './components/NewsletterView'
 
 import './NewsletterRoute.css'
