+++ src/components/NewsletterListItem.js
